main/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
whitenoise==6.0.0
django-heroku==0.3.1
drf-yasg==1.20.0
